refactor(ocr): use generic call lifecycle wrapper
