+++ litellm-rust/crates/ai-gateway/src/lib.rs
@@ -3,15 +3,16 @@
 //! Two layers, split by feature so the Python `cdylib` can depend on the I/O
 //! without pulling in the HTTP server:
 //!
-//! - [`io`]: all network I/O (OCR HTTP call, realtime WebSocket splice, the
-//!   pre-warmed realtime pool). Always available — no feature required. The
-//!   Python bridge links this for `run_ocr`.
+//! - Call-type modules such as [`ocr`]: provider transforms, lifecycle hooks,
+//!   and provider I/O. Always available — no feature required.
+//! - [`io`]: compatibility exports and realtime WebSocket splice helpers.
 //! - The server modules ([`auth`], [`routes`], [`state`]) and anything pulling
 //!   `axum` are gated behind the `server` feature, which the `litellm-ai-gateway`
 //!   binary turns on. The `python-config` feature additionally pulls in [`python`]
 //!   for the load-time config reader.
 
 pub mod io;
+pub mod ocr;
 
 /// GIL-activity tracking. Pure (atomics only); shared by the `server` routes and
 /// the `python-config` reader, so it is available without either feature.
