+++ litellm-rust/crates/ai-gateway/src/ocr/client.rs
@@ -1,0 +1,14 @@
+use std::sync::OnceLock;
+use std::time::Duration;
+
+const OCR_TIMEOUT_SECS: u64 = 600;
+
+pub(super) fn http_client() -> &'static reqwest::Client {
+    static CLIENT: OnceLock<reqwest::Client> = OnceLock::new();
+    CLIENT.get_or_init(|| {
+        reqwest::Client::builder()
+            .timeout(Duration::from_secs(OCR_TIMEOUT_SECS))
+            .build()
+            .expect("failed to build reqwest client")
+    })
+}
